Meet Your Match Update - First Pass
In essence, all this does is remove the main menu and gamemenu background files temporarily and enables the new versioning system the TF team has implemented. This commit also includes the beginnings of my conversion to the #base method, which will allow hud updates to be handled much more manageably. There will 100% be more updates as I find out what is and isn't broken.