8 October 2017 - Health Appearance Changes
These changes affect how the TargetID appears. Instead of a cross next to a teammate's name, you will now see a colored box behind the person's health. This overall takes up less space, and I feel is an overall improvement. I'll have to weigh the benefits of adding the original as a customization option, as it would require a lot of tinkering.