Fix TargetID team bar on different aspect ratios
Pushing this to development first because I haven't tested it much at all, and want to give it a few playthroughs in-game to make sure it isn't borked. I also removed the engineer's "pick up building" menu (and thus the TargetID extender) because I was running into too many issues with them.