Materials clean up
Mostly removing VMT parameters that don't affect anything. Also swapped in some $color parameters into a few VMTs to reduce number of VTFs needed

For curious hud developers: $vertexcolor is needed for any button that makes use of `image_drawcolor` / `image_armedcolor` to change the color of the VTF. $vertexalpha is required if you want to use any amonut of partial transparency in the alpha for `image_drawcolor` / `image_armedcolor.

I left both parameters in and set to 1 for all VMTs just for consistency.