add border to disconnect/quit dialogs
going to do a half-assed 12am explanation because I'm guessing i'll forget what even happened here

there are two problem aspects to this, and two panels tied to them:

- `ConfirmDialog`, which doesn't allow you to add your own border
- `ExplanationLabel`, which has some inconsistent positioning behavior and adjusts dynamically

The tall value of ExplanationLabel is to some extent hard-coded, because it is used to dictate the size of ConfirmDialog (the overall panel). If your font gets bigger, the whole ConfirmDialog gets better. In addition to this, multiple hud elements use the confirmdialog.res file. Which means...if you want to add your own border, you can add your own custom element and hide the ConfirmDialog border (with `"paintborder" "0"`), but your element will never be the correct size because ConfirmDialog's tall dynamically adjusts, and yours is static. So your disconnect dialog in casual will look different than the quit dialog (or, a better example, the "coach players" dialog) because they have varying amounts of text

another facet of this is that ExplanationLabel behaves..strangely? it's center wrapped by default, and long story short you SHOULD be able to ypos cs-0.5 a centerwrapped element but you can't in this case idk). so instead, I kept the original ExplanationLabel so it could dictate the overall size of ConfirmDialog still but I just made the text invisible. I then added my own ExplanationLabel that would center a bit more sensibly (because not being able to adjust the ypos with a cs value was absolutely owning me). then from there, I had to adjust some of the disconnect abandon panels so they'd be a little bit better centered as well, but that was easy since my custom ExplanationLabel could actually have its ypos properly adjusted

Now to go back to adding the borders. I added my own ImagePanel and set it to cs-0.5/cs-0.5/f0/f0. I noticed that for some reason when I set proportionaltoparent to 0, two of the four borders would draw. if I set it to 1, three of the four borders would draw. I'm guessing it has something to do with the dynamically adjusting tall of ConfirmDialog, but I don't know. doesn't matter, now I'm basically 75% of the way to it looking how I want it to.

I honestly don't know why, maybe they're just coded this way...but the ypos of the misc confirm/quit/etc buttons would always appropriately move with the bottom of ConfirmDialog, so I never had to worry about buttons getting cut off. with that, I decided to solve the "missing bottom border" issue by just straight up faking it and adding an imagepanel that was pinned to one of these buttons and then offset to be in the correct border position.

everything works on my end in this moment, but I'm sure I'll realize something is broken at some point. i also don't know how much this explanation will help me in the future, probably will forget what 90% of it even means. I guess we'll see if it works lmao